+++ src/database/repositories/faqs.repository.ts
@@ -1,12 +1,7 @@
-import { IDatabase, IMain } from 'pg-promise';
 import { BaseRepository } from './base.repository';
 import { CreateFaqSchema, Faq, UpdateFaqSchema } from '../models';
 
 export class FaqsRepository extends BaseRepository {
-    constructor(protected readonly db: IDatabase<any>, protected readonly pgp: IMain) {
-        super(db, pgp);
-    }
-
     getAllFaqsForListing(listingId: string): Promise<Faq[]> {
         return this.db.manyOrNone('SELECT * FROM faq WHERE listing_id = $1', listingId);
     }