+++ src/jobs/dbBackup.job.ts
@@ -8,11 +8,10 @@ import { S3ClientService, s3ClientService } from '../services/s3Client.service';
 class DbBackupJob extends BaseJob {
     constructor(private readonly s3ClientService: S3ClientService) {
         super();
-        this.s3ClientService = s3ClientService;
     }
 
     get frequency() {
-        return '* 0 5 * * *';
+        return '0 0 5 * * *';
     }
 
     get command() {